Conflicts:
	cots/org.jboss.cache/.project
	edexOsgi/build.edex/esb/conf/spring/project.properties
	edexOsgi/com.raytheon.edex.plugin.text/src/com/raytheon/edex/plugin/text/dao/TextDao.java
	edexOsgi/com.raytheon.uf.edex.decodertools/src/com/raytheon/uf/edex/decodertools/bufr/BUFRSection1.java
	edexOsgi/com.raytheon.uf.edex.decodertools/src/com/raytheon/uf/edex/decodertools/time/TimeTools.java
	javaUtilities/com.raytheon.wes2bridge.utility.product/.project
	ncep/gov.noaa.nws.ncep.common.dataplugin.sgwh/src/gov/noaa/nws/ncep/common/dataplugin/sgwh/SgwhRecord.java
	ncep/gov.noaa.nws.ncep.common.dataplugin.sgwhv/src/gov/noaa/nws/ncep/common/dataplugin/sgwhv/SgwhvRecord.java
	ncep/gov.noaa.nws.ncep.common.dataplugin.ssha/src/gov/noaa/nws/ncep/common/dataplugin/ssha/SshaRecord.java
	ncep/gov.noaa.nws.ncep.common.staticdata/.project

Former-commit-id: 975053c3dcd3014ca1bdc06d46193420b896d2ca [formerly c7024992a32f78fea0f68351aa042ab9ad856473] [formerly 37793c86b188c287fc51ce1b39b0e7728f55bc08] [formerly 975053c3dcd3014ca1bdc06d46193420b896d2ca [formerly c7024992a32f78fea0f68351aa042ab9ad856473] [formerly 37793c86b188c287fc51ce1b39b0e7728f55bc08] [formerly d325d496f50f682d586e8fe099dd0de3f798d388 [formerly 37793c86b188c287fc51ce1b39b0e7728f55bc08 [formerly a435e6464a78345b5036e3060430375dc53e4faf]]]]
Former-commit-id: d325d496f50f682d586e8fe099dd0de3f798d388
Former-commit-id: 5e0e2d2c48e51f5fa8ae8c297f03deaa0e786188 [formerly 23c34c469a49140b5d7862855195aeb104ad241b] [formerly 788e16ce64aaeca91000aa565d43e89a27246362 [formerly cf3c768213d486d7e726b6655c0d246224ec63fa]]
Former-commit-id: bbdcec83196c1a64fd2e31c1e9abdcf46698ab0b [formerly 3dd5b84aa2dacc20a15694b092eb40ae7fd2a581]
Former-commit-id: c01b0962427c1567f61257d75ca87723a2176c9b

+++ edexOsgi/com.raytheon.edex.plugin.text/src/com/raytheon/edex/plugin/text/dao/TextDao.java
@@ -51,24 +51,31 @@ import com.raytheon.uf.edex.database.query.DatabaseQuery;
  */
 public class TextDao extends DefaultPluginDao {
 
-    public TextDao(String pluginName) throws PluginException {
-        super(pluginName);
-    }
+	public TextDao(String pluginName) throws PluginException {
+		super(pluginName);
+	}
+
+	@Override
+	public void purgeAllData() {
+		logger.warn("purgeAllPluginData not implemented for text. No data will be purged.");
+	}
+
+	protected void loadScripts() throws PluginException {
+		// no op
+	}
 
     @Override
-    public void purgeAllData() {
-        logger.warn("purgeAllPluginData not implemented for text. No data will be purged.");
-    }
+	public void purgeExpiredData() throws PluginException {
+		int deletedRecords = 0;
 
-    protected void loadScripts() throws PluginException {
-        // no op
-    }
+		// only do full purge every few hours since incremental purge runs every
+		// minute
+		if (Calendar.getInstance().get(Calendar.HOUR_OF_DAY) % 3 == 0) {
+			TextDB.purgeStdTextProducts();
+		}
 
-    @Override
-    public void purgeExpiredData() throws PluginException {
-        int deletedRecords = TextDB.purgeStdTextProducts();
-        PurgeLogger.logInfo("Purged " + deletedRecords + " items total.",
-                "text");
+		PurgeLogger.logInfo("Purged " + deletedRecords + " items total.",
+				"text");
     }
 
     @SuppressWarnings("unchecked")
@@ -106,5 +113,5 @@ public class TextDao extends DefaultPluginDao {
         } else {
             return result.get(0).getTime();
         }
-    }
+	}
 }
